% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{bootRisk}
\alias{bootRisk}
\title{bootRisk}
\usage{
bootRisk(control, dat, thetal0, mix = c(1, 0))
}
\arguments{
\item{control}{Object computed by `computeControl()`}

\item{dat}{Object computed by `computeData()`}

\item{thetal0}{Estimated \eqn{\theta}{θ} from \eqn{\lambda^{(k)}}{λ⁽ᵏ⁾} for use in bootstrap}

\item{mix}{Uses Gramian mix[1]*G + mix[2]*M, which corresponds to
a weighted average of risks in f-space and h-space,
respectively}
}
\value{
`logR` is a function for which logR(l) returns the
    estimated MISE of the estimator with penalty 10^l
}
\description{
Computes estimated risk function
}
\author{
David Kent
}
