% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{chooseLambda}
\alias{chooseLambda}
\title{chooseLambda}
\usage{
chooseLambda(control, dat, eps = 1e-04, maxit = 100, ll0 = -1, verbose = FALSE)
}
\arguments{
\item{control}{Object computed by `computeControl()`}

\item{dat}{Object computed by `computeData()`}

\item{eps}{Convergence criterion}

\item{maxit}{Maximum number of iterations}

\item{ll0}{Initial \eqn{\lambda}{λ}}

\item{verbose}{Output information at each iteration}
}
\value{
\eqn{\lambda}{λ} chosen by iterated parametric bootstrap
}
\description{
Chooses lambda
}
\author{
David Kent
}
