% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{computeConstraints}
\alias{computeConstraints}
\title{computeConstraints}
\usage{
computeConstraints(B, nxeval = 1000)
}
\arguments{
\item{B}{The B-spline basis}

\item{nxeval}{Number of points x at which to compute B(x)}
}
\value{

}
\description{
Computes constraints for quadratic programming problem.
}
\details{
In either the constrained quadratic program, or the metric
projection, the PDF and positive constraints are required by
`quadprog` to be expressed as linear functions of the
coefficients, and those matrices and vectors are computed
here. Since the positivity constraint is only approximate, we need
the B-spline basis and a grid density on which to express the
constraint.
}
\author{
David Kent
}
