% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{computeControl}
\alias{computeControl}
\title{computeControl}
\usage{
computeControl(Y, K, qdim, gtwid, padperc)
}
\arguments{
\item{Y}{The data}

\item{K}{The number of histogram bins}

\item{qdim}{The dimension of the spline space}

\item{gtwid}{The Fourier transform of the error density}

\item{padperc}{How far beyond the data the spline support should
extend, as a proportion of the range of Y}
}
\value{

}
\description{
Computes all the details required for the estimator
}
\details{
Utility function which computes the histogram grid, B-spline
basis, inner- and cross-product matrices, and linear constraint
matrices.

The objects computed here do *not* depend on the data except that
the data is used to determine the intervals on which the histogram
and spline basis are computed; thus, if new data is acquired
(e.g. through simulations), which are consistent with those
intervals, these objects may all be re-used.
}
\author{
David Kent
}
