% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{computeData}
\alias{computeData}
\title{computeData}
\usage{
computeData(Y, control)
}
\arguments{
\item{Y}{The (possibly new) data}

\item{control}{A `control` object from computeControl()}
}
\value{
A list containing two objects: `$hhat` is the histogram bin-heights, and `$Y` is the data.
}
\description{
(Re-)computes `hhat` on the `midp` grid in `control`
}
\author{
David Kent
}
