% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{computeHist}
\alias{computeHist}
\title{computeHist}
\usage{
computeHist(Y, K, pad = NULL, padperc = 0)
}
\arguments{
\item{Y}{The data; a numeric vector}

\item{K}{The number of bin midpoints}

\item{pad}{Distance (in units of Y) beyond [min(Y),max(Y)] to compute bins}

\item{padperc}{Similar to `pad`, but expressed as a percentage of the range of Y}
}
\value{

}
\description{
Computes bin midpoints and densities.
}
\details{
Padding is not typically useful, but in circumstances where new
data will be used with the same `control` object
(e.g. simulations), some padding can help avoid the necessity of
recomputing the histogram bins and associated inner product
matrices.
}
\author{
David Kent
}
