% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolutions.R
\name{computeM}
\alias{computeM}
\title{Computes Gram matrix of convolved B-splines}
\usage{
computeM(gtwid, qdim, knots, h)
}
\arguments{
\item{gtwid}{Fourier transform of g}

\item{qdim}{Dimension of spline basis}

\item{knots}{Location of knots}

\item{h}{Knot spacing}
}
\value{

}
\description{
computeM
}
\details{
Computes \eqn{q \times q}{q × q} matrix M with entries \eqn{M_{ij} = \int (g*B_i)(g*B_j)}{Mᵢⱼ = ∫(g*Bᵢ)(g*Bⱼ)}
}
\author{
David Kent
}
