% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolutions.R
\name{computeN}
\alias{computeN}
\title{computeN}
\usage{
computeN(basis, midp, gtwid, K, qdim)
}
\arguments{
\item{basis}{The B-spline basis}

\item{midp}{The K midpoints of the histogram}

\item{gtwid}{The Fourier transform of the error density}

\item{K}{The number of histogram midpoints}

\item{qdim}{The dimension of the B-Spline basis}
}
\value{

}
\description{
Computes the matrix N for the quadratic programming problem
}
\details{
Computes \eqn{K\times q}{K × q} matrix with entries
\deqn{N_{ij} = \int_a^b g*B_j,}{Nᵢⱼ = ∫ₐᵇg*Bⱼ,}
where \eqn{a = m_i-\delta/2}{a = mᵢ-δ/2} and \eqn{b = m_i+\delta/2}{b = mᵢ+δ/2}, with \eqn{m_i}{mᵢ} the midpoint of the ith histogram bin.
}
\author{
David Kent
}
