% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{decon}
\alias{decon}
\title{decon}
\usage{
decon(Y,gtwid=NULL,lambda="risk",K=NULL,qdim=NULL,
    control=NULL,padperc=0.1,unconstrained=FALSE,
    metric=TRUE,verbose=FALSE)
}
\arguments{
\item{Y}{The data, i.e. Y = X + e}

\item{gtwid}{A function which returns the Fourier transform of the density of e}

\item{lambda}{Either "risk" to choose a penalty parameter automatically, or a positive number}

\item{K}{(optional) number of bins for histogram}

\item{qdim}{(optional) dimension of spline space}

\item{control}{(optional) control object from previous call to decon()}

\item{padperc}{(optional) How far beyond the data the spline
support should extend, as a proportion of the range of Y}

\item{unconstrained}{(optional) If TRUE, returns unconstrained
solutionmetric (optional) If TRUE, projects
unconstrained solution onto PDFs. If FALSE, computes solution
as constrained minimization problem.}

\item{metric}{Compute metric projection? If FALSE, adds constraint directly to minimization problem}

\item{verbose}{(optional) If TRUE, prints extra information.}
}
\value{
A list, where `$fhat` is a function which evaluates the
    estimate of the density of X, `$dat` contains the data used to
    compute that estimate, and `$control` contains the
    miscellaneous objects used to compute `$fhat`. `$control` can
    be supplied to another call of decon() so that it does not
    have to be recomputed.
}
\description{
Deconvolves a density
}
\details{
Given data Y = X + e, where e has density g, computes an estimate
of the density of X. The Fourier transform of the density of e
must be provided as `gtwid`.
}
\author{
David Kent
}
