% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{doSolve}
\alias{doSolve}
\title{doSolve}
\usage{
doSolve(control, dat, lambda, unconstrained = FALSE, metric = TRUE)
}
\arguments{
\item{control}{Control object produced by computeControl()}

\item{dat}{Data object produced by computeData()}

\item{lambda}{Penalty parameter}

\item{unconstrained}{If TRUE, do not apply PDF constraints}

\item{metric}{If TRUE, compute metric projection of unconstrained
solution; otherwise, apply constraint directly in minimization
problem.}
}
\value{

}
\description{
Does solve, given `control`, `dat`, and `lambda`
}
\details{
With all the pre-computed pieces from computeControl() and
computeData(), solve the minimization problem over the spline
space.
}
\author{
David Kent
}
