% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{getgtwid}
\alias{getgtwid}
\title{getgtwid()}
\usage{
getgtwid(g = "Gaussian", param)
}
\arguments{
\item{g}{Character string, one of "Gaussian", "Laplace", or
"Uniform"}

\item{param}{Named numeric vector containing an entry named 's'
for the standard deviation of a Gaussian, 'b' for the scale of
the Laplace, or 'a' for the interval half-width for the
Uniform.}
}
\value{
Function which evaluates the Fourier transform of the
    specified density
}
\description{
Returns Fourier Transform of specified error density
}
\details{
For Gaussian density \eqn{g(x) = \sqrt{2\pi s^2}\exp{-x^2/2s^2}}{g(x) = √(2πs²)·exp{ x²/2s² }}, returns
\deqn{F[g](t) = \int e^{-itx} g(x)\,dx = \exp{-s^2t^2/2}}{F[g](t) = ∫e⁻ⁱᵗˣg(x)dx = exp(-s²t²/2)}

For Laplace density \eqn{g(x) = \exp{-|x|/b}/2b}{g(x) = exp{ -|x|/b }/2b}, returns
\deqn{F[g](t) = \int e^{-itx} g(x)\,dx = 1/(1+b^2t^2)}{F[g](t) = ∫e⁻ⁱᵗˣg(x)dx = 1/(1+b²t²)}

For Uniform density \eqn{g(x) = I[-a,a]/2a}{g(x) = Ind[-a,a]/2a}, returns
\deqn{F[g](t) = \int e^{-itx}g(x)\,dx = sinc(at)}{F[g](t) = ∫e⁻ⁱᵗˣg(x)dx = sinc(at)}
}
\author{
David Kent
}
