% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{histogtwid}
\alias{histogtwid}
\title{Histogram Basis Function Fourier Transform}
\usage{
histogtwid(t, midp, delt)
}
\arguments{
\item{t}{Argument of Fourier transform}

\item{midp}{Midpoint of histogram bin}

\item{delt}{Width of histogram bin}
}
\value{
(complex-valued) Fourier transform of histogram basis function evaluated at t
}
\description{
Fourier transform of histogram basis function
 
The basis functions of a histogram are indicator functions for the
bins, scaled by the bin-spacing delt. This function evaluates the
Fourier transform of the basis function at midpoint `midp` with
spacing `delt`.
}
