% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{unscaledtwid}
\alias{unscaledtwid}
\title{Unscaled B-Spline Fourier Transform}
\usage{
unscaledtwid(t)
}
\arguments{
\item{t}{Argument of Fourier Transform}
}
\value{
(complex-valued) Fourier transform of B_{0,4} evaluated at t
}
\description{
Unscaled B-Spline Fourier Transform
}
\details{
Given fourth-order integer-knot B-Splines, evaluates the Fourier transform of B_{0,4}.
}
